feat(Algorithm): Initial version of ACO

+++ src/main/java/com/deliverif/app/models/algorithms/astar/Scorer.java
@@ -1,5 +1,5 @@
 package com.deliverif.app.models.algorithms.astar;
 
 public interface Scorer<T extends GraphNode> {
-    double computeCost(T from, T to);
+    float computeCost(T from, T to);
 }
